include/code_loading.hrl: New header to help with code loading
It provides a macro called `load_nif_from_host_triple_dir/2` to help
with handling the loop which tries all native code dirs one by one,
until it can load the library (or not).

This is macro instead of a regular module/function because the
`erlang:load_nif/2` must be called from the Erlang module corresponding
to the NIF being loaded.